# feat(typescript-estree): switch to globby